extract binary/docker starting logic into helper along with the simulated fuel node
