+++ Cargo.toml
@@ -2,9 +2,10 @@
 resolver = "2"
 members = [
   "committer",
-  "e2e",
+  "e2e/benches",
+  "e2e/helpers",
+  "e2e/tests",
   "fee_algo_simulation",
-  "fuel_simulator",
   "packages/adapters/clock",
   "packages/adapters/eth",
   "packages/adapters/fuel",
@@ -28,16 +29,19 @@ rust-version = "1.82"
 publish = false
 
 [workspace.dependencies]
-async-graphql = { version = "7", default-features = false }
-async-graphql-warp = { version = "7", default-features = false }
 clock = { path = "./packages/adapters/clock", default-features = false }
 eth = { path = "./packages/adapters/eth", default-features = false }
 fuel = { path = "./packages/adapters/fuel", default-features = false }
-storage = { path = "./packages/adapters/storage", default-features = false }
+fuel-block-committer-encoding = { path = "./packages/encoding", default-features = false }
 metrics = { path = "./packages/metrics", default-features = false }
 services = { path = "./packages/services", default-features = false }
+storage = { path = "./packages/adapters/storage", default-features = false }
 test-helpers = { path = "./packages/test-helpers/", default-features = false }
-fuel-block-committer-encoding = { path = "./packages/encoding", default-features = false }
+e2e-helpers = { path = "./e2e/helpers", default-features = false }
+
+async-graphql = { version = "7", default-features = false }
+env_logger = { version = "0.11", default-features = false }
+async-graphql-warp = { version = "7", default-features = false }
 actix-web = { version = "4", default-features = false }
 bitvec = { version = "1.0", default-features = false }
 bytesize = { version = "1.3", default-features = false }
