+++ e2e/benches/src/main.rs
@@ -1,7 +1,4 @@
-mod server;
-
 use clap::Parser;
-use server::run_server;
 
 #[derive(Parser, Debug)]
 #[command(author, version, about, long_about = None)]
@@ -13,16 +10,15 @@ struct Args {
     port: u16,
 }
 
-fn main() {
+#[tokio::main]
+async fn main() -> anyhow::Result<()> {
     if std::env::var("RUST_LOG").is_err() {
         std::env::set_var("RUST_LOG", "debug");
     }
 
     env_logger::init();
     let args = Args::parse();
-    tokio::runtime::Builder::new_current_thread()
-        .enable_all()
-        .build()
-        .unwrap()
-        .block_on(run_server(args.block_size, args.port));
+    e2e_helpers::fuel_node_simulated::run_server(args.block_size, args.port).await;
+
+    Ok(())
 }
