+++ e2e/helpers/Cargo.toml
@@ -1,13 +1,14 @@
 [package]
-name = "e2e"
-authors = { workspace = true }
-edition = { workspace = true }
-homepage = { workspace = true }
-license = { workspace = true }
-repository = { workspace = true }
-version = { workspace = true }
-publish = { workspace = true }
-rust-version = { workspace = true }
+name = "e2e-helpers"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+homepage.workspace = true
+license.workspace = true
+repository.workspace = true
+rust-version.workspace = true
+publish.workspace = true
+
 build = "build.rs"
 
 [build-dependencies]
@@ -24,7 +25,7 @@ tokio = { workspace = true, features = [
 walkdir = { workspace = true }
 zip = { workspace = true, features = ["deflate"] }
 
-[dev-dependencies]
+[dependencies]
 alloy = { workspace = true, features = [
   "signer-aws",
   "signer-mnemonic",
@@ -61,3 +62,8 @@ tokio = { workspace = true, features = [
   "fs",
 ] }
 url = { workspace = true }
+warp = { workspace = true }
+async-graphql = { workspace = true, features = ["playground"] }
+async-graphql-warp = { workspace = true }
+clap = { workspace = true, features = ["derive", "std"] }
+env_logger = { version = "0.11.6" }
