+++ e2e/helpers/src/fuel_node.rs
