+++ e2e/helpers/src/fuel_node_simulated/server.rs
@@ -1,0 +1,2 @@
+
+
