+++ e2e/tests/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "e2e-tests"
+authors = { workspace = true }
+edition = { workspace = true }
+homepage = { workspace = true }
+license = { workspace = true }
+repository = { workspace = true }
+version = { workspace = true }
+publish = { workspace = true }
+rust-version = { workspace = true }
+
+[dev-dependencies]
+e2e-helpers = { workspace = true }
+anyhow = { workspace = true }
+tokio = { workspace = true, features = ["macros", "rt-multi-thread"] }
+storage = { workspace = true }
+services = { workspace = true }
