--- fuel_simulator/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "fuel_simulator"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-tokio = { workspace = true, features = ["full"] }
-warp = { workspace = true }
-async-graphql = { workspace = true, features = ["playground"] }
-async-graphql-warp = { workspace = true }
-clap = { workspace = true, features = ["derive"] }
-env_logger = { version = "0.11.6" }
-hex = { workspace = true, features = ["alloc"] }
-rand = { workspace = true }
